Axios query with CORS-Anywhere working
